backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
agno
openai
agent-sandbox
python-dotenv
sqlalchemy
sqlalchemy
